feat: BaseAtKeysIo to bring in enc/dec/gen of keys
